ea - added test for profits that brought up test coverage and deleted test for profits table because that ended up being included in profits
