--- frontend/src/fixtures/profitsFixtures.js
@@ -1,11 +1,0 @@
-const profitsFixtures = {
-    oneProfit:
-        [
-            {
-                "profit": 10,
-                "date": "06/11/2021"
-            }
-        ]
-}
-
-export default profitsFixtures;
